Add support for database update scripts.
The files /docker-entrypoint-updatedb.d/* are run every time a container starts, even if the database directory is not empty. This is unlike the files /docker-entrypoint-initdb.d/*, which are only run iif the database directory is empty. When the database directory is empty, the update scripts are run after the init scripts.

Suppose that the following files are present in a container image:
/docker-entrypoint-initdb.d/init.sh
/docker-entrypoint-updatedb.d/update.sh

When a new container is run for the first time with a persistent volume, for example as follows:

$ docker run --rm -it \
	-v ./my-initdb.d:/docker-entrypoint-initdb.d \
        -v ./my-updatedb.d:/docker-entrypoint-updatedb.d \
	-v my-vol:/var/lib/postgresql/data postgres

the following init files are run, in this order:

/docker-entrypoint-initdb.d/init.sh
/docker-entrypoint-updatedb.d/update.sh

If the container is stopped and a new container is started with the same command, only this file is run:

/docker-entrypoint-updatedb.d/update.sh