Add liburing support to 18+
This is a bit complicated to use in Docker (because `io_uring` has a long and storied security story), but it is a new feature of 18+ that we should probably explicitly support for users who want to jump through the hoops to use it despite the security risk.

Additionally, this re-alphabetizes our `--with-xxx` flags.